- using the C version of the scheduler for now - changed system calls to use C++11

+++ src/system.h
@@ -1,7 +1,9 @@
 #ifndef SYSTEM_H_
 #define SYSTEM_H_
 
-long sys_get_tick_count(void);
-long sys_get_cpu_count(void);
+#include "def.h"
+
+int64 sys_get_tick_count(void);
+int32 sys_get_cpu_count(void);
 
 #endif //SYSTEM_H_
